refactor tutorial steps with requirement checks
